backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "GPL-V3.0"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.2"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pylint = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
